+++ src/interfaces/index.ts
@@ -1,3 +1,4 @@
 export * from './auth_token';
 export * from './client_options';
 export * from './refresh_access_token_options';
+export * from './passwordless_verify_response';