+++ CHANGELOG.md
@@ -6,3 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- `passwordlessLogin` method to `BrainClient` which starts passworldess login flow by sending confirmation e-mail
+- `verifyPasswordlessLogin` method to `BrainClient` which checks status of passwordless login